requirements.txt
# Core data science libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Machine learning
lightgbm>=3.3.0
optuna>=3.0.0

# Feature engineering
ta>=0.10.0
PyWavelets>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.60.0

# CLI and configuration
typer>=0.7.0
pyyaml>=6.0
colorama>=0.4.0

# Optional: For advanced config management (uncomment if using)
# hydra-core>=1.3.0

# Optional: For experiment tracking (uncomment if using)
# mlflow>=2.0.0
# wandb>=0.13.0 